+++ service-module/src/main/java/com/gmail/evgenymoshchin/service/impl/ReviewServiceImpl.java
@@ -4,6 +4,7 @@ import com.gmail.evgenymoshchin.repository.ReviewRepository;
 import com.gmail.evgenymoshchin.repository.model.Review;
 import com.gmail.evgenymoshchin.service.ReviewService;
 import com.gmail.evgenymoshchin.service.converters.ReviewServiceConverter;
+import com.gmail.evgenymoshchin.service.exception.ServiceException;
 import com.gmail.evgenymoshchin.service.model.ReviewDTO;
 import com.gmail.evgenymoshchin.service.model.ReviewPageDTO;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +15,11 @@ import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 
+import static com.gmail.evgenymoshchin.service.constants.ExceptionConstants.REVIEW_WAS_NOT_FOUND_MESSAGE;
+import static com.gmail.evgenymoshchin.service.constants.ExceptionConstants.VISIBILITY_CHANGE_MESSAGE;
+import static com.gmail.evgenymoshchin.service.constants.ExceptionConstants.VISIBILITY_WILL_CHANGE;
 import static com.gmail.evgenymoshchin.service.util.ServiceUtil.getNumbersOfPages;
 
 @Log4j2
@@ -47,7 +52,11 @@ public class ReviewServiceImpl implements ReviewService {
     @Transactional
     public void removeById(Long id) {
         Review review = reviewRepository.findById(id);
-        reviewRepository.remove(review);
+        if (Objects.nonNull(review)) {
+            reviewRepository.remove(review);
+        } else {
+            throw new ServiceException(String.format(REVIEW_WAS_NOT_FOUND_MESSAGE, id));
+        }
     }
 
     @Override
@@ -62,20 +71,32 @@ public class ReviewServiceImpl implements ReviewService {
         }
         for (Long id : visibleIdsFromDB) {
             if (!selectedIds.contains(id)) {
-                log.info("This review id {}, wasn't contain in visibleIdsFromDB list, visibility will change", id);
+                log.info(VISIBILITY_CHANGE_MESSAGE, id);
                 findReviewByIdUpdateVisible(id);
             }
         }
         for (Long id : selectedIds) {
             if (!visibleIdsFromDB.contains(id)) {
-                log.info("This review id {}, wasn't contain in selectedIds list, visibility will change", id);
+                log.info(VISIBILITY_WILL_CHANGE, id);
                 findReviewByIdUpdateVisible(id);
             }
         }
     }
 
+    @Override
+    @Transactional
+    public ReviewDTO addReview(ReviewDTO reviewDTO, String username) {
+        Review review = reviewServiceConverter.convertDTOToReview(reviewDTO, username);
+        reviewRepository.persist(review);
+        return reviewServiceConverter.convertReviewToDTO(review);
+    }
+
     private void findReviewByIdUpdateVisible(Long id) {
         Review review = reviewRepository.findById(id);
-        review.setIsVisible(!review.getIsVisible());
+        if (Objects.nonNull(review)) {
+            review.setIsVisible(!review.getIsVisible());
+        } else {
+            throw new ServiceException(String.format(REVIEW_WAS_NOT_FOUND_MESSAGE, id));
+        }
     }
 }
